Organização dos arquivos do projeto.
