--- libpythonpro_exercicio/tests/test_spam/conftest.py
@@ -1,20 +1,0 @@
-import pytest
-
-from libpythonpro_exercicio.spam.db import Conexao
-
-
-@pytest.fixture()
-def sessao(conexao):
-    sessao_obj = conexao.gerar_sessao()
-    yield sessao_obj
-    sessao_obj.roll_back()
-    sessao_obj.fechar()
-
-
-@pytest.fixture(scope='session')
-def conexao():
-    # Setup
-    conexao_obj = Conexao()
-    yield conexao_obj
-    # Tear Down
-    conexao_obj.fechar()
