+++ setup.py
@@ -75,7 +75,7 @@ def find_package_data(
                         break
                 if bad_name:
                     continue
-                if (os.path.isfile(os.path.join(fn, "libpythonpro_exercicio/__init__.py"))
+                if (os.path.isfile(os.path.join(fn, "libpythonproexercicio/__init__.py"))
                         and not prefix):
                     if not package:
                         new_package = name
@@ -101,7 +101,7 @@ def find_package_data(
     return out
 
 
-PACKAGE = "libpythonpro_exercicio"
+PACKAGE = "libpythonproexercicio"
 NAME = PACKAGE
 DESCRIPTION = "Módulo para exemplificar construção de projetos Python no curso PyTools"
 AUTHOR = "Renzo Nuccitelli"
